give role via user, users can actions btn with their permission, users can edit their profile and logout alert with ajax

+++ app/Http/Controllers/DepartmentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreDepartment;
 use App\Http\Requests\UpdateDepartment;
 use App\Models\Department;
+use App\Models\User;
 use Illuminate\Validation\Rule;
 use Yajra\DataTables\Facades\DataTables;
 
@@ -17,6 +18,10 @@ class DepartmentController extends Controller
 
     public function index()
     {
+        if (!User::findOrFail(auth()->id())->can('view_departments')) {
+            return abort(401);
+        }
+
         if (\request()->ajax()) {
             $data = Department::latest()->get();
             return DataTables::of($data)
@@ -26,9 +31,16 @@ class DepartmentController extends Controller
                     return $actionBtn;
                 })
                 ->addColumn('actions', function ($row) {
-                    $edit = '<a href="' . route("departments.edit", $row->id) . '" class="btn btn-sm btn-warning">' . '<i class="fa-solid fa-edit"></i>' . '</a>';
+                    $edit = '';
+                    $delete = '';
 
-                    $delete = '<a href="#" id="del-btn" data-id="' . $row->id . '" class="btn btn-sm btn-danger">' . '<i class="fa-solid fa-trash-alt"></i>' . '</a>';
+                    if (User::findOrFail(auth()->id())->can('edit_department')) {
+                        $edit = '<a href="' . route("departments.edit", $row->id) . '" class="btn btn-sm btn-warning">' . '<i class="fa-solid fa-edit"></i>' . '</a>';
+                    }
+
+                    if (User::findOrFail(auth()->id())->can('remove_department')) {
+                        $delete = '<a href="#" id="del-btn" data-id="' . $row->id . '" class="btn btn-sm btn-danger">' . '<i class="fa-solid fa-trash-alt"></i>' . '</a>';
+                    }
 
                     return "<div class='btn-group'>$edit$delete</div>";
                 })
@@ -40,11 +52,19 @@ class DepartmentController extends Controller
 
     public function create()
     {
+        if (!User::findOrFail(auth()->id())->can('create_department')) {
+            return abort(401);
+        }
+
         return view('department.create');
     }
 
     public function store(StoreDepartment $request)
     {
+        if (!User::findOrFail(auth()->id())->can('create_department')) {
+            return abort(401);
+        }
+
         $department = new Department;
         $department->name = $request->name;
         $department->save();
@@ -54,6 +74,10 @@ class DepartmentController extends Controller
 
     public function edit($id)
     {
+        if (!User::findOrFail(auth()->id())->can('edit_department')) {
+            return abort(401);
+        }
+
         $department = Department::findOrFail($id);
         return view('department.edit', [
             'department' => $department,
@@ -62,6 +86,10 @@ class DepartmentController extends Controller
 
     public function update($id, UpdateDepartment $request)
     {
+        if (!User::findOrFail(auth()->id())->can('edit_department')) {
+            return abort(401);
+        }
+
         request()->validate([
             "name" => [Rule::unique('departments', 'name')->ignore($id)]
         ]);
@@ -75,6 +103,10 @@ class DepartmentController extends Controller
 
     public function destroy($id)
     {
+        if (!User::findOrFail(auth()->id())->can('remove_department')) {
+            return abort(401);
+        }
+
         $department = Department::findOrFail($id);
         $department->delete();
