+++ app/Http/Controllers/EmployeeController.php
@@ -9,6 +9,7 @@ use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
+use Spatie\Permission\Models\Role;
 use Yajra\DataTables\Facades\DataTables;
 
 class EmployeeController extends Controller
@@ -20,6 +21,10 @@ class EmployeeController extends Controller
 
     public function index()
     {
+        if (!User::findOrFail(auth()->id())->can('view_employees')) {
+            return abort(401);
+        }
+
         if (\request()->ajax()) {
             $data = User::with('department')->latest()->get();
             return DataTables::of($data)
@@ -29,16 +34,34 @@ class EmployeeController extends Controller
                     return $actionBtn;
                 })
                 ->addColumn('actions', function ($row) {
-                    $edit = '<a href="' . route("employees.edit", $row->id) . '" class="btn btn-sm btn-warning">' . '<i class="fa-solid fa-edit"></i>' . '</a>';
+                    $edit = '';
+                    $detail = '';
+                    $delete = '';
 
-                    $detail = '<a href="' . route("employees.show", $row->id) . '" class="btn btn-sm btn-outline-secondary">' . '<i class="fa-solid fa-circle-info"></i>' . '</a>';
+                    if (User::findOrFail(auth()->id())->can('edit_employee')) {
+                        $edit = '<a href="' . route("employees.edit", $row->id) . '" class="btn btn-sm btn-warning">' . '<i class="fa-solid fa-edit"></i>' . '</a>';
+                    }
 
-                    $delete = '<a href="#" id="del-btn" data-id="' . $row->id . '" class="btn btn-sm btn-danger">' . '<i class="fa-solid fa-trash-alt"></i>' . '</a>';
+                    if (User::findOrFail(auth()->id())->can('view_employees')) {
+                        $detail = '<a href="' . route("employees.show", $row->id) . '" class="btn btn-sm btn-outline-secondary">' . '<i class="fa-solid fa-circle-info"></i>' . '</a>';
+                    }
+
+                    if (User::findOrFail(auth()->id())->can('remove_employee')) {
+                        $delete = '<a href="#" id="del-btn" data-id="' . $row->id . '" class="btn btn-sm btn-danger">' . '<i class="fa-solid fa-trash-alt"></i>' . '</a>';
+                    }
 
                     return "<div class='btn-group'>$edit$detail$delete</div>";
                 })
                 ->addColumn('department', function ($row) {
                     return $row->department ? $row->department->name : '-';
+                })
+                ->addColumn('roles', function ($row) {
+                    $lists = '';
+                    foreach ($row->roles as $role) {
+                        $lists .= "<span class='badge bg-primary m-1'>" . $role->name . "</span>";
+                    }
+                    $listsContainer = "<div class='d-flex justify-content-center flex-wrap'>$lists</div>";
+                    return $listsContainer;
                 })
                 ->editColumn('avatar', function ($row) {
                     return $row->avatar ? "<img style='width: 80px;' src='/storage/$row->avatar'/>" : '-';
@@ -46,7 +69,7 @@ class EmployeeController extends Controller
                 ->editColumn('is_present', function ($row) {
                     return $row->is_present ? "<span class='badge bg-primary'>Yes</span>" : "<span class='badge bg-danger'>Leave</span>";
                 })
-                ->rawColumns(['action', 'actions', 'avatar', 'is_present'])
+                ->rawColumns(['action', 'actions', 'avatar', 'is_present', 'roles'])
                 ->make(true);
         }
         return view('employee.index');
@@ -54,6 +77,10 @@ class EmployeeController extends Controller
 
     public function show($id)
     {
+        if (!User::findOrFail(auth()->id())->can('view_employees')) {
+            return abort(401);
+        }
+
         $employee = User::findOrFail($id);
 
         return view("employee.show", ["employee" => $employee]);
@@ -61,13 +88,22 @@ class EmployeeController extends Controller
 
     public function create()
     {
+        if (!User::findOrFail(auth()->id())->can('create_employee')) {
+            return abort(401);
+        }
+
         return view('employee.create', [
             'departments' => Department::all(),
+            'roles' => Role::all(),
         ]);
     }
 
     public function store(StoreEmployee $request)
     {
+        if (!User::findOrFail(auth()->id())->can('create_employee')) {
+            return abort(401);
+        }
+
         $employee = new User;
         $employee->employee_id = $request->employee_id;
         $employee->name = $request->name;
@@ -86,20 +122,33 @@ class EmployeeController extends Controller
         }
         $employee->save();
 
+        $employee->syncRoles($request->roles);
+
         return redirect(route('employees.index'))->with('created', 'New Employee Created Successful');
     }
 
     public function edit($id)
     {
+        if (!User::findOrFail(auth()->id())->can('edit_employee')) {
+            return abort(401);
+        }
+
         $employee = User::findOrFail($id);
+
         return view('employee.edit', [
             'employee' => $employee,
             'departments' => Department::all(),
+            'roles' => Role::all(),
+            'old_roles' => $employee->roles()->pluck('id')->toArray(),
         ]);
     }
 
     public function update($id, UpdateEmployee $request)
     {
+        if (!User::findOrFail(auth()->id())->can('edit_employee')) {
+            return abort(401);
+        }
+
         request()->validate([
             "employee_id" => Rule::unique('users', 'employee_id')->ignore($id),
             "email" => Rule::unique('users', 'email')->ignore($id),
@@ -138,6 +187,8 @@ class EmployeeController extends Controller
 
         $employee->update();
 
+        $employee->syncRoles($request->roles);
+
         if (auth()->id() == $id && $request->password) {
             Auth::guard('web')->logout();
             $request->session()->invalidate();
@@ -150,6 +201,10 @@ class EmployeeController extends Controller
 
     public function destroy($id)
     {
+        if (!User::findOrFail(auth()->id())->can('remove_employee')) {
+            return abort(401);
+        }
+
         $employee = User::findOrFail($id);
         $employee->delete();
