+++ app/Http/Controllers/RoleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreRole;
 use App\Http\Requests\UpdateRole;
+use App\Models\User;
 use Illuminate\Validation\Rule;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
@@ -18,6 +19,10 @@ class RoleController extends Controller
 
     public function index()
     {
+        if (!User::findOrFail(auth()->id())->can('view_roles')) {
+            return abort(401);
+        }
+
         if (\request()->ajax()) {
             $data = Role::latest()->get();
             return DataTables::of($data)
@@ -27,9 +32,16 @@ class RoleController extends Controller
                     return $actionBtn;
                 })
                 ->addColumn('actions', function ($row) {
-                    $edit = '<a href="' . route("roles.edit", $row->id) . '" class="btn btn-sm btn-warning">' . '<i class="fa-solid fa-edit"></i>' . '</a>';
+                    $edit = '';
+                    $delete = '';
 
-                    $delete = '<a href="#" id="del-btn" data-id="' . $row->id . '" class="btn btn-sm btn-danger">' . '<i class="fa-solid fa-trash-alt"></i>' . '</a>';
+                    if (User::findOrFail(auth()->id())->can('edit_role')) {
+                        $edit = '<a href="' . route("roles.edit", $row->id) . '" class="btn btn-sm btn-warning">' . '<i class="fa-solid fa-edit"></i>' . '</a>';
+                    }
+
+                    if (User::findOrFail(auth()->id())->can('remove_role')) {
+                        $delete = '<a href="#" id="del-btn" data-id="' . $row->id . '" class="btn btn-sm btn-danger">' . '<i class="fa-solid fa-trash-alt"></i>' . '</a>';
+                    }
 
                     return "<div class='btn-group'>$edit$delete</div>";
                 })
@@ -38,7 +50,9 @@ class RoleController extends Controller
                     foreach ($row->permissions as $permission) {
                         $lists .= "<span class='badge bg-primary m-1'>" . $permission->name . "</span>";
                     }
-                    return $lists;
+                    $listsContainer = "<div class='d-flex flex-wrap'>$lists</div>";
+
+                    return $listsContainer;
                 })
                 ->rawColumns(['action', 'actions', 'permissions'])
                 ->make(true);
@@ -48,11 +62,19 @@ class RoleController extends Controller
 
     public function create()
     {
+        if (!User::findOrFail(auth()->id())->can('create_role')) {
+            return abort(401);
+        }
+
         return view('role.create', ['permissions' => Permission::all()]);
     }
 
     public function store(StoreRole $request)
     {
+        if (!User::findOrFail(auth()->id())->can('create_role')) {
+            return abort(401);
+        }
+
         $role = new Role;
         $role->name = $request->name;
         $role->save();
@@ -64,6 +86,10 @@ class RoleController extends Controller
 
     public function edit($id)
     {
+        if (!User::findOrFail(auth()->id())->can('edit_role')) {
+            return abort(401);
+        }
+
         $role = Role::findOrFail($id);
         return view('role.edit', [
             'role' => $role,
@@ -74,6 +100,10 @@ class RoleController extends Controller
 
     public function update($id, UpdateRole $request)
     {
+        if (!User::findOrFail(auth()->id())->can('edit_role')) {
+            return abort(401);
+        }
+
         request()->validate([
             "name" => [Rule::unique('roles', 'name')->ignore($id)]
         ]);
@@ -91,6 +121,10 @@ class RoleController extends Controller
 
     public function destroy($id)
     {
+        if (!User::findOrFail(auth()->id())->can('remove_role')) {
+            return abort(401);
+        }
+
         $role = Role::findOrFail($id);
         $role->delete();
